+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - #678 Added logging when an exception is thrown when charon is being deleted by moodle
 - #681 updated composer/composer to 2.1.9
 - #567 Activity completion for charon
+- #636 If charon does not have defences, the defence icon would not be shown submission list
 
 ## [1.5.4] - 2021-10-11
 
@@ -34,7 +35,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - #638 Added css so that when tester returns styled output, it will be displayed that way
 - #589 Added the ability for teachers to add feedback for students' submissions
 - #635 Backup and restore possibility for charon_review_comment table
-- #636 If charon does not have defences, the defence icon would not be shown submission list
 
 ## [1.5.3] - 2021-10-08