+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 ### Added
+- #676 Student has an option to toggle current submission comments and all comments for this charon. Same for the teacher.
 - #589 When a teacher adds a new review comment about a student's submission, an eye symbol on the same submission will turn 
   red on the student's view. When submission is opened by the student, the red color will be removed again.
 - #678 Added logging when an exception is thrown when charon is being deleted by moodle