+++ CHANGELOG.md
@@ -4,9 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
-
-### Fixed
+## [1.5.5] - 2021-10-22
 
 ### Added
 - #589 When a teacher adds a new review comment about a student's submission, an eye symbol on the same submission will turn